+++ packages/frontend/src/components/DataViz/store/lineChartSlice.ts
@@ -1,5 +1,6 @@
 import { ChartKind, isVizLineChartConfig } from '@lightdash/common';
 import { createSlice } from '@reduxjs/toolkit';
+import { prepareAndFetchChartData } from '../../../features/sqlRunner/store/thunks';
 import {
     resetChartState,
     setChartConfig,
@@ -14,6 +15,20 @@ export const lineChartConfigSlice = createSlice({
         ...cartesianChartConfigSlice.caseReducers,
     },
     extraReducers: (builder) => {
+        // Include the extraReducers from cartesianChartConfigSlice
+        builder.addCase(prepareAndFetchChartData.pending, (state) => {
+            state.chartDataLoading = true;
+            state.chartDataError = undefined;
+        });
+        builder.addCase(prepareAndFetchChartData.fulfilled, (state, action) => {
+            state.chartDataLoading = false;
+            state.series = action.payload?.valuesColumns;
+            state.chartData = action.payload;
+        });
+        builder.addCase(prepareAndFetchChartData.rejected, (state, action) => {
+            state.chartDataLoading = false;
+            state.chartDataError = new Error(action.error.message);
+        });
         builder.addCase(setChartOptionsAndConfig, (state, action) => {
             if (action.payload.type !== ChartKind.LINE) {
                 return;
@@ -37,9 +52,10 @@ export const lineChartConfigSlice = createSlice({
                 state.display = action.payload.display;
             }
         });
-        builder.addCase(resetChartState, () =>
-            cartesianChartConfigSlice.getInitialState(),
-        );
+        builder.addCase(resetChartState, () => ({
+            ...cartesianChartConfigSlice.getInitialState(),
+            type: ChartKind.LINE,
+        }));
     },
 });