+++ packages/frontend/src/pages/SemanticViewerEdit.tsx
@@ -2,6 +2,7 @@ import { subject } from '@casl/ability';
 import { useEffect, useMemo } from 'react';
 import { Provider } from 'react-redux';
 import { useHistory, useParams, useRouteMatch } from 'react-router-dom';
+import { useUnmount } from 'react-use';
 import Page from '../components/common/Page/Page';
 import { setChartOptionsAndConfig } from '../components/DataViz/store/actions/commonChartActions';
 import { getChartConfigAndOptions } from '../components/DataViz/transformers/getChartConfigAndOptions';
@@ -16,6 +17,7 @@ import { SemanticViewerResultsRunnerFrontend } from '../features/semanticViewer/
 import { selectSemanticViewerState } from '../features/semanticViewer/store/selectors';
 import {
     initializeSemanticViewer,
+    resetState,
     SemanticViewerStateStatus,
 } from '../features/semanticViewer/store/semanticViewerSlice';
 import { store } from '../features/sqlRunner/store';
@@ -184,6 +186,10 @@ const SemanticViewerEditorPageWithStore = () => {
         chartResultsQuery.isSuccess,
         chartResultsQuery.data,
     ]);
+
+    useUnmount(() => {
+        dispatch(resetState());
+    });
 
     // TODO: add error state
     if (